+++ lib/src/us/kbase/kbballgown/BallgownInput.java
@@ -2,6 +2,7 @@
 package us.kbase.kbballgown;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import javax.annotation.Generated;
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
@@ -22,8 +23,9 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
  * num_threads: number of threads
  * workspace_name: the name of the workspace it gets saved to
  * optional params:
+ * run_all_combinations: run all paired condition combinations
+ * condition_labels: conditions for expression set object
  * maximum_num_genes: used to filter genes in the differential expression matrix
- * fold_scale_type: one of ["linear", "log2+1", "log10+1"]
  * </pre>
  * 
  */
@@ -32,90 +34,57 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 @JsonPropertyOrder({
     "expressionset_ref",
     "diff_expression_matrix_set_suffix",
-    "alpha_cutoff",
-    "fold_change_cutoff",
     "num_threads",
     "workspace_name",
-    "maximum_num_genes",
-    "fold_scale_type"
+    "input_type",
+    "run_all_combinations",
+    "condition_labels"
 })
 public class BallgownInput {
 
     @JsonProperty("expressionset_ref")
-    private String expressionsetRef;
+    private java.lang.String expressionsetRef;
     @JsonProperty("diff_expression_matrix_set_suffix")
-    private String diffExpressionMatrixSetSuffix;
-    @JsonProperty("alpha_cutoff")
-    private Double alphaCutoff;
-    @JsonProperty("fold_change_cutoff")
-    private Double foldChangeCutoff;
+    private java.lang.String diffExpressionMatrixSetSuffix;
     @JsonProperty("num_threads")
     private Long numThreads;
     @JsonProperty("workspace_name")
-    private String workspaceName;
-    @JsonProperty("maximum_num_genes")
-    private Long maximumNumGenes;
-    @JsonProperty("fold_scale_type")
-    private String foldScaleType;
-    private Map<String, Object> additionalProperties = new HashMap<String, Object>();
+    private java.lang.String workspaceName;
+    @JsonProperty("input_type")
+    private java.lang.String inputType;
+    @JsonProperty("run_all_combinations")
+    private Long runAllCombinations;
+    @JsonProperty("condition_labels")
+    private List<String> conditionLabels;
+    private Map<java.lang.String, Object> additionalProperties = new HashMap<java.lang.String, Object>();
 
     @JsonProperty("expressionset_ref")
-    public String getExpressionsetRef() {
+    public java.lang.String getExpressionsetRef() {
         return expressionsetRef;
     }
 
     @JsonProperty("expressionset_ref")
-    public void setExpressionsetRef(String expressionsetRef) {
+    public void setExpressionsetRef(java.lang.String expressionsetRef) {
         this.expressionsetRef = expressionsetRef;
     }
 
-    public BallgownInput withExpressionsetRef(String expressionsetRef) {
+    public BallgownInput withExpressionsetRef(java.lang.String expressionsetRef) {
         this.expressionsetRef = expressionsetRef;
         return this;
     }
 
     @JsonProperty("diff_expression_matrix_set_suffix")
-    public String getDiffExpressionMatrixSetSuffix() {
+    public java.lang.String getDiffExpressionMatrixSetSuffix() {
         return diffExpressionMatrixSetSuffix;
     }
 
     @JsonProperty("diff_expression_matrix_set_suffix")
-    public void setDiffExpressionMatrixSetSuffix(String diffExpressionMatrixSetSuffix) {
+    public void setDiffExpressionMatrixSetSuffix(java.lang.String diffExpressionMatrixSetSuffix) {
         this.diffExpressionMatrixSetSuffix = diffExpressionMatrixSetSuffix;
     }
 
-    public BallgownInput withDiffExpressionMatrixSetSuffix(String diffExpressionMatrixSetSuffix) {
+    public BallgownInput withDiffExpressionMatrixSetSuffix(java.lang.String diffExpressionMatrixSetSuffix) {
         this.diffExpressionMatrixSetSuffix = diffExpressionMatrixSetSuffix;
-        return this;
-    }
-
-    @JsonProperty("alpha_cutoff")
-    public Double getAlphaCutoff() {
-        return alphaCutoff;
-    }
-
-    @JsonProperty("alpha_cutoff")
-    public void setAlphaCutoff(Double alphaCutoff) {
-        this.alphaCutoff = alphaCutoff;
-    }
-
-    public BallgownInput withAlphaCutoff(Double alphaCutoff) {
-        this.alphaCutoff = alphaCutoff;
-        return this;
-    }
-
-    @JsonProperty("fold_change_cutoff")
-    public Double getFoldChangeCutoff() {
-        return foldChangeCutoff;
-    }
-
-    @JsonProperty("fold_change_cutoff")
-    public void setFoldChangeCutoff(Double foldChangeCutoff) {
-        this.foldChangeCutoff = foldChangeCutoff;
-    }
-
-    public BallgownInput withFoldChangeCutoff(Double foldChangeCutoff) {
-        this.foldChangeCutoff = foldChangeCutoff;
         return this;
     }
 
@@ -135,63 +104,78 @@ public class BallgownInput {
     }
 
     @JsonProperty("workspace_name")
-    public String getWorkspaceName() {
+    public java.lang.String getWorkspaceName() {
         return workspaceName;
     }
 
     @JsonProperty("workspace_name")
-    public void setWorkspaceName(String workspaceName) {
+    public void setWorkspaceName(java.lang.String workspaceName) {
         this.workspaceName = workspaceName;
     }
 
-    public BallgownInput withWorkspaceName(String workspaceName) {
+    public BallgownInput withWorkspaceName(java.lang.String workspaceName) {
         this.workspaceName = workspaceName;
         return this;
     }
 
-    @JsonProperty("maximum_num_genes")
-    public Long getMaximumNumGenes() {
-        return maximumNumGenes;
+    @JsonProperty("input_type")
+    public java.lang.String getInputType() {
+        return inputType;
     }
 
-    @JsonProperty("maximum_num_genes")
-    public void setMaximumNumGenes(Long maximumNumGenes) {
-        this.maximumNumGenes = maximumNumGenes;
+    @JsonProperty("input_type")
+    public void setInputType(java.lang.String inputType) {
+        this.inputType = inputType;
     }
 
-    public BallgownInput withMaximumNumGenes(Long maximumNumGenes) {
-        this.maximumNumGenes = maximumNumGenes;
+    public BallgownInput withInputType(java.lang.String inputType) {
+        this.inputType = inputType;
         return this;
     }
 
-    @JsonProperty("fold_scale_type")
-    public String getFoldScaleType() {
-        return foldScaleType;
+    @JsonProperty("run_all_combinations")
+    public Long getRunAllCombinations() {
+        return runAllCombinations;
     }
 
-    @JsonProperty("fold_scale_type")
-    public void setFoldScaleType(String foldScaleType) {
-        this.foldScaleType = foldScaleType;
+    @JsonProperty("run_all_combinations")
+    public void setRunAllCombinations(Long runAllCombinations) {
+        this.runAllCombinations = runAllCombinations;
     }
 
-    public BallgownInput withFoldScaleType(String foldScaleType) {
-        this.foldScaleType = foldScaleType;
+    public BallgownInput withRunAllCombinations(Long runAllCombinations) {
+        this.runAllCombinations = runAllCombinations;
+        return this;
+    }
+
+    @JsonProperty("condition_labels")
+    public List<String> getConditionLabels() {
+        return conditionLabels;
+    }
+
+    @JsonProperty("condition_labels")
+    public void setConditionLabels(List<String> conditionLabels) {
+        this.conditionLabels = conditionLabels;
+    }
+
+    public BallgownInput withConditionLabels(List<String> conditionLabels) {
+        this.conditionLabels = conditionLabels;
         return this;
     }
 
     @JsonAnyGetter
-    public Map<String, Object> getAdditionalProperties() {
+    public Map<java.lang.String, Object> getAdditionalProperties() {
         return this.additionalProperties;
     }
 
     @JsonAnySetter
-    public void setAdditionalProperties(String name, Object value) {
+    public void setAdditionalProperties(java.lang.String name, Object value) {
         this.additionalProperties.put(name, value);
     }
 
     @Override
-    public String toString() {
-        return ((((((((((((((((((("BallgownInput"+" [expressionsetRef=")+ expressionsetRef)+", diffExpressionMatrixSetSuffix=")+ diffExpressionMatrixSetSuffix)+", alphaCutoff=")+ alphaCutoff)+", foldChangeCutoff=")+ foldChangeCutoff)+", numThreads=")+ numThreads)+", workspaceName=")+ workspaceName)+", maximumNumGenes=")+ maximumNumGenes)+", foldScaleType=")+ foldScaleType)+", additionalProperties=")+ additionalProperties)+"]");
+    public java.lang.String toString() {
+        return ((((((((((((((((("BallgownInput"+" [expressionsetRef=")+ expressionsetRef)+", diffExpressionMatrixSetSuffix=")+ diffExpressionMatrixSetSuffix)+", numThreads=")+ numThreads)+", workspaceName=")+ workspaceName)+", inputType=")+ inputType)+", runAllCombinations=")+ runAllCombinations)+", conditionLabels=")+ conditionLabels)+", additionalProperties=")+ additionalProperties)+"]");
     }
 
 }